xelors update
Updated. i added the other engine i was making. i ended up not going with LODs for it because the polys are on the lower side. but please let me know if the power to fuel consumption ration needs adjustment or do it yourself because i am not good at balancing game items.